Refactor widget views into separate files
- Extract CoinPriceView into its own file
- Remove inline CoinPriceRow from PriceWidgetView
- Follow single view per file pattern
